pdf2png/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
certifi==2024.2.2
charset-normalizer==3.3.2
gptscript @ git+https://github.com/gptscript-ai/py-gptscript.git@0cebee3afa51b8c56006e479a990c262bd693ba6
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
idna==3.6
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.14
PyMuPDFb==1.24.10
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.12.2
urllib3==2.2.1
